New sliding window for input rate estimation (jobs_avg) and stateful producers
NOTE: work in progress! Compiles, but not yet fully working.

The idea with stateful producers is that you can specify a producer behaviour
{Mod,Args}, where the default is {jobs_prod_simple, F}. When dispatching new
jobs from the producer queue, Mod:next(...) -> {Fun, NewState} is called.

The sliding window function in jobs_avg is intended to estimate input rate
in multiple different time windows:
- the most recent time slot ("burst estimation")
- a sliding-window average (avg over the kept time slots)
- an exponential sum calculated on the sliding-window average
The idea is to have an efficient calculation that gives a decent basis for
determining whether to approve a request immediately or to queue instead.
The simplest thing is to always queue, but this imposes a latency penalty on
all jobs which not all (read: Basho) are willing to accept. Also, the current
rate-based regulation is limited in how fast it can go (due to timer
resolution, queue/dequeue overhead, etc).